rtnl: change l_rtnl_address expire times to uint64_t
These are calculated from uint64_t values so they also should be
uint64_t sizes.

+++ ell/rtnl.c
@@ -54,8 +54,8 @@ struct l_rtnl_address {
 	char label[IFNAMSIZ];
 	uint32_t preferred_lifetime;
 	uint32_t valid_lifetime;
-	uint32_t preferred_expiry_time;
-	uint32_t valid_expiry_time;
+	uint64_t preferred_expiry_time;
+	uint64_t valid_expiry_time;
 	uint32_t flags;
 };
